possible fix for bar charts after r51022
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51964 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/gui/dialog/options/model/FillingModel.java
@@ -207,7 +207,7 @@ public class FillingModel extends MultipleOptionsModel {
 		double alpha = alpha0;
 		int idx = getFillingListener().getSelectedBarIndex();
 		if (idx != 0) {
-			float barAlpha = algo.getBarAlpha(idx);
+			double barAlpha = algo.getBarAlpha(idx);
 			if (barAlpha != -1) {
 				alpha = barAlpha;
 			}
